[rust] rearrange code, split code into modules. We support D1-like SoC by now.
