+++ rust/src/soc.rs
@@ -1,0 +1,1 @@
+pub mod d1;
